Proto changes

+++ app/schema/types/Split.scala
@@ -1,16 +1,16 @@
 package schema.types
 
-import com.types.{Split => protoSplit}
+import com.assetmantle.schema.types.base.{Split => protoSplit}
 import schema.id.OwnableID
 import schema.id.base.IdentityID
 
-case class Split(ownerID: IdentityID, ownableID: OwnableID, value: BigDecimal) {
+case class Split(ownerID: IdentityID, ownableID: OwnableID, value: BigInt) {
 
-  def send(out: BigDecimal): Split = this.copy(value = this.value - out)
+  def send(out: BigInt): Split = this.copy(value = this.value - out)
 
-  def receive(in: BigDecimal): Split = this.copy(value = this.value + in)
+  def receive(in: BigInt): Split = this.copy(value = this.value + in)
 
-  def canSend(out: BigDecimal): Boolean = this.value >= out
+  def canSend(out: BigInt): Boolean = this.value >= out
 
   def asProtoSplit: protoSplit = protoSplit.newBuilder().setOwnerID(ownerID.asProtoIdentityID).setOwnableID(ownableID.toAnyOwnableID).setValue(value.toString()).build()
 
@@ -20,7 +20,7 @@ case class Split(ownerID: IdentityID, ownableID: OwnableID, value: BigDecimal) {
 
 object Split {
 
-  def apply(split: protoSplit): Split = Split(ownerID = IdentityID(split.getOwnerID), ownableID = OwnableID(split.getOwnableID), value = BigDecimal(split.getValue))
+  def apply(split: protoSplit): Split = Split(ownerID = IdentityID(split.getOwnerID), ownableID = OwnableID(split.getOwnableID), value = split.getValue.toLong)
 
   def apply(protoBytes: Array[Byte]): Split = Split(protoSplit.parseFrom(protoBytes))
